+++ tests/testthat/test-map.R
@@ -39,10 +39,6 @@ test_that("types automatically coerced upwards", {
   expect_identical(map_chr(as.raw(0:255), identity), as.character(as.raw(0:255)))
 })
 
-test_that("map_raw",{
-  expect_equal(map_raw("a", charToRaw), charToRaw("a"))
-})
-
 test_that("logical and integer NA become correct double NA", {
   expect_identical(
     map_dbl(list(NA, NA_integer_), identity),
@@ -170,5 +166,4 @@ test_that("map() with empty input copies names", {
   expect_identical(map_int(named_list, identity), named(int()))
   expect_identical(map_dbl(named_list, identity), named(dbl()))
   expect_identical(map_chr(named_list, identity), named(chr()))
-  expect_identical(map_raw(named_list, identity), named(raw()))
 })